# Optimize model-init runtime